pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scribe-cli"
version = "0.1.0"
description = "CLI para transcripción de audio con detección de speakers y timestamps"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Ivan Landaverde" }
]
keywords = ["transcription", "speech-to-text", "whisper", "diarization", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "openai-whisper>=20231117",
    "simple-diarizer>=0.0.15",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pydub>=0.25.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
scribe = "scribe.cli:main"

[project.urls]
Homepage = "https://github.com/ivanlandaverde/scribe"
Repository = "https://github.com/ivanlandaverde/scribe"
Issues = "https://github.com/ivanlandaverde/scribe/issues"

[tool.hatch.build.targets.wheel]
packages = ["scribe"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
